Reform of page numbering
Pages are numbered starting with the title page, so physical page
numbering matches logical (except if you typeset hardcover, which
is another page 1).

We let LaTeX do its magic and enable typesetting of page numbers inside
chapters. This magically makes blank pages generated by \openright
numberless, which is right. But it also means that it is very hard
to avoid page numbers in the table of contents.

Also, we no longer add a blank page at the very end of the thesis
to make the number of pages even.